src/requirements.txt
Django==1.11.29
image==1.5.32
Pillow==6.2.2
pytz==2020.1
six==1.15.0
appdirs==1.4.4  
certifi==2016.2.28  
django==1.11.29  
enum==0.4.7  
glm==0.4.3  
image 1.5.32  
numpy==1.16.6  
packaging==20.9  
pip==9.0.1  
pyopengl==3.1.5  
pyopengl-accelerate==3.1.5  
pyparsing==2.4.7  
pypiwin32==223  
python==2.7.13  
pytz==2020.1  
pywin32==228  
setuptools==36.4.0  
vc==9  
vs2008_runtime==9.00.30729.5054  
wheel==0.29.0  
wincertstore==0.2